requirements-python3.10.txt
abcmeta==2.2.1
contextlib2==21.6.0
flake8==7.1.1
keras==2.15.0
keras-tuner==1.4.7
kfp==1.8.22
mergedeep==1.3.4
mypy==1.13.0
numpy==1.26.4
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
pre_commit==4.0.1
pytest==8.3.3
pytest-cov==6.0.0
PyYAML==6.0.2
tensorflow==2.15.1
tensorflow-datasets==4.9.7
tfx==1.15.1
types-PyYAML==6.0.12
